pyproject.toml
[tool.poetry]
name = "joyofenergy"
version = "0.1.5"
description = ""
authors = [
    "Albert Attard <albertattard@gmail.com>",
    "Baran Demir <social@bdemir.de>",
    "Björn Eberhardt <beberhar@thoughtworks.com>",
    "Dan Lafeir <danlafeir@users.noreply.github.com>",
    "Ina Hartmann <nhartman@thoughtworks.com>",
    "Jaydeep Chakrabarty <jaydeepc@in-Jaydeepc-3089.local>",
    "Martin Soto <donsoto@gmail.com>",
    "Nilakhya Chatterjee <nilakhya.chatterjee@thoughtworks.com>",
]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
fastapi = "^0.110.1"
httpx = "^0.27.0"
loguru = "^0.7.2"
python = "^3.12"
uvicorn = "^0.29.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.tests]
optional = true
[tool.poetry.group.tests.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"

[tool.poetry.group.ci]
optional = true
[tool.poetry.group.ci.dependencies]
ruff = "^0.3.5"
black = "^24.3.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
lint.select = ["I00", "E", "F", "W"]

[tool.black]
line-length = 120
target-version = ['py39']